+++ setup.cfg
@@ -29,7 +29,8 @@ cpc-sbom = templates/*.jinja2
 
 [options.entry_points]
 console_scripts = 
-	cpc-sbom = cpc_sbom:generate_sbom
+	cpc-sbom = cpc_sbom.generate:generate_sbom
+	cpc-sbom-inject = cpc_sbom.inject:inject_artifact_name
 
 [bumpversion:file:setup.py]
 search = version="{current_version}"
